+++ scim/src/main/java/org/cloudfoundry/identity/uaa/scim/dao/common/ScimGroupMembershipManager.java
@@ -15,8 +15,8 @@ package org.cloudfoundry.identity.uaa.scim.dao.common;
 import java.util.List;
 import java.util.Set;
 
+import org.cloudfoundry.identity.uaa.scim.domain.common.ScimGroupInterface;
 import org.cloudfoundry.identity.uaa.scim.domain.common.ScimGroupMemberInterface;
-import org.cloudfoundry.identity.uaa.scim.domain.standard.ScimGroupInterface;
 import org.cloudfoundry.identity.uaa.scim.exception.MemberAlreadyExistsException;
 import org.cloudfoundry.identity.uaa.scim.exception.MemberNotFoundException;
 import org.cloudfoundry.identity.uaa.scim.exception.ScimResourceNotFoundException;
